requirements.txt
django
django-guardian
django-phonenumber-field
phonenumbers
django-autocomplete-light
django-bootstrap-modal-forms